pyproject.toml
[project]
name = "claude-telegram-bot"
version = "0.1.0"
description = "Generic Telegram bot interface powered by Claude Code CLI"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "python-dotenv>=1.0",
    "typer>=0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["*.py"]
exclude = ["__pycache__"]
